refactor: relocate MockWebSocket to frontend/test-utils
